requirements.txt
Flask==2.0.2
openai==0.19.0
pandas==1.3.4
progressbar33==2.4
scikit_learn==1.2.2
